services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
orjson==3.10.7
paho-mqtt==2.1.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
